Deduplicate `tt_001` `CreateEventEntity` function
