+++ src/servant/tt_000/45E4.c
@@ -3,31 +3,7 @@
 #include <sfx.h>
 #include <psxsdk/libc.h>
 
-void CreateEventEntity(Entity* entityParent, s32 entityId, s32 params) {
-    Entity* entity;
-    s32 i;
-
-    for (i = 0; i < 3; i++) {
-        entity = &g_Entities[5 + i];
-        if (!entity->entityId) {
-            break;
-        }
-    }
-
-    if (!entity->entityId) {
-        DestroyEntity(entity);
-        entity->entityId = entityId;
-        entity->zPriority = entityParent->zPriority;
-        entity->facingLeft = entityParent->facingLeft;
-        entity->flags = FLAG_KEEP_ALIVE_OFFCAMERA;
-        entity->posX.val = entityParent->posX.val;
-        entity->posY.val = entityParent->posY.val;
-        // Not necessarily making batFamBlueTrail here, but
-        // that's an Ext that works. Just needs parent at 0x8C.
-        entity->ext.batFamBlueTrail.parent = entityParent;
-        entity->params = params;
-    }
-}
+#include "../create_event_entity.h"
 
 s32 func_801746A0(s32 arg0) {
     if (PLAYER.velocityY < 0 && !(g_Player.pl_vram_flag & 1)) {
