+++ src/servant/tt_001/F84.c
@@ -67,7 +67,7 @@ INCLUDE_ASM("servant/tt_001/nonmatchings/F84", func_us_80172D10);
 
 #include "../process_event.h"
 
-INCLUDE_ASM("servant/tt_001/nonmatchings/F84", CreateEventEntity);
+#include "../create_event_entity.h"
 
 INCLUDE_ASM("servant/tt_001/nonmatchings/F84", func_us_80173378);
